// src/enums/coreEnums.ts
export enum EventEnums {
  pressKey = "pressKey",
  clickButton = "clickButton",
  mouseClick = "mouseClick",
}

export enum TypeEnums {
  userEvent = "userEvent",
}

export enum StateEnums {
  isOpenBuildMenu = "isOpenBuildMenu",
  isPreviewMode = "isPreviewMode",
}

export enum BildMenuButtonEnums {
  Button = "Button",
  Image = "Image",
  Video = "Video",
  Text = "Text",
  Canvas = "Canvas",
  Line = "Line",
  Color = "Color",
}